test: share Doctor persistence fixture contexts
