+++ src/commands/doctor-config-flow.workspace-persistence.test.ts
@@ -1,6 +1,6 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it } from "vitest";
 import { resolveAgentWorkspaceDir } from "../agents/agent-scope-config.js";
 import { readConfigFileSnapshot } from "../config/config.js";
 import { withEnvOverride, withTempHome, writeOpenClawConfig } from "../config/test-helpers.js";
@@ -11,44 +11,12 @@ import {
   runInitialConfigWriteHealth,
   runWriteConfigHealth,
 } from "../flows/doctor-health-contribution-runners.config.js";
-import type { DoctorHealthFlowContext } from "../flows/doctor-health-contribution-types.js";
-import type { RuntimeEnv } from "../runtime.js";
 import {
   closeOpenClawStateDatabaseForTest,
   openOpenClawStateDatabase,
 } from "../state/openclaw-state-db.js";
-import { loadAndMaybeMigrateDoctorConfig } from "./doctor-config-flow.js";
-import { createDoctorPrompter, type DoctorOptions } from "./doctor-prompter.js";
+import { prepareDoctorContext } from "./doctor-config-flow.test-support.js";
 import { normalizeCompatibilityConfigValues } from "./doctor/shared/legacy-config-core-migrate.js";
-
-async function prepareDoctorContext(configPath: string): Promise<DoctorHealthFlowContext> {
-  const runtime: RuntimeEnv = { error: vi.fn(), exit: vi.fn(), log: vi.fn() };
-  const options: DoctorOptions = { nonInteractive: true, repair: true };
-  const prompter = createDoctorPrompter({ runtime, options });
-  const configResult = await loadAndMaybeMigrateDoctorConfig({
-    options,
-    confirm: (params) => prompter.confirm(params),
-    runtime,
-    prompter,
-  });
-  return {
-    runtime,
-    options,
-    prompter,
-    configResult,
-    cfg: configResult.cfg,
-    cfgForPersistence: structuredClone(configResult.cfg),
-    sourceConfigValid: configResult.sourceConfigValid ?? true,
-    configPath,
-    stateDirExistedAtStart: true,
-    ...(configResult.runWithPluginMetadataSnapshot
-      ? { runWithPluginMetadataSnapshot: configResult.runWithPluginMetadataSnapshot }
-      : {}),
-    ...(configResult.invalidatePluginMetadataSnapshot
-      ? { invalidatePluginMetadataSnapshot: configResult.invalidatePluginMetadataSnapshot }
-      : {}),
-  };
-}
 
 describe("Doctor workspace persistence", () => {
   afterEach(() => {
